updates package-lock after conflicts, moved nav code to jsx
